Refactored codegen a lot so we could store line numbers alongside bytecode. Dissassembler now prints code lines alongside the bytecode. Re-implemented and and or operators to skip unnecessary computations

+++ source/polyscript/lexer.h
@@ -65,7 +65,7 @@ struct Token {
     size_t m_length { 0 };
 
     char* m_pLineStart { nullptr };
-    size_t m_line { 0 };
+    uint32_t m_line { 0 };
 };
 
 ResizableArray<Token> Tokenize(IAllocator* pAllocator, String sourceText);
